+++ genomics_data_index/test/integration/storage/index/test_KmerIndexer.py
@@ -4,7 +4,7 @@ from pathlib import Path
 from sourmash import load_file_as_signatures
 
 from genomics_data_index.test.integration import data_dir
-from genomics_data_index.test.integration.variant.index import get_values_from_signatures
+from genomics_data_index.test.integration.storage.index import get_values_from_signatures
 from genomics_data_index.storage.index.KmerIndexer import KmerIndexerSourmash
