+++ genomics_data_index/test/integration/storage/io/test_NucleotideSampleDataSequenceMask.py
@@ -5,7 +5,7 @@ from typing import Dict
 import pytest
 
 from genomics_data_index.test.integration import regular_vcf_dir, data_dir
-from genomics_data_index.test.integration.variant.io import read_vcf_df
+from genomics_data_index.test.integration.storage.io import read_vcf_df
 from genomics_data_index.storage.io.mutation.NucleotideSampleDataSequenceMask import NucleotideSampleDataSequenceMask
